requirements.txt
reflex>=0.4.7
openai>=1.14.0
python-dotenv==1.0.1
